feat(Web-UI): Added the quality of the episode in the tv show detail overview
