+++ src/WebAPI/ClientApp/src/lang/babel_edit.babel
@@ -2198,6 +2198,29 @@
 												</translations>
 											</concept_node>
 											<concept_node>
+												<name>quality</name>
+												<description/>
+												<comment/>
+												<translations>
+													<translation>
+														<language>de-DE</language>
+														<approved>false</approved>
+													</translation>
+													<translation>
+														<language>en-US</language>
+														<approved>false</approved>
+													</translation>
+													<translation>
+														<language>fr-FR</language>
+														<approved>false</approved>
+													</translation>
+													<translation>
+														<language>pl-PL</language>
+														<approved>false</approved>
+													</translation>
+												</translations>
+											</concept_node>
+											<concept_node>
 												<name>title</name>
 												<description/>
 												<comment/>
