+++ src/WebAPI/ClientApp/src/lang/pl-PL.json
@@ -164,7 +164,8 @@
                 "media-size": "Rozmiar",
                 "added-at": "Dodano w",
                 "updated-at": "Zaktualizowano w",
-                "actions": "Działania"
+                "actions": "Działania",
+                "quality": "Jakość"
             }
         },
         "server-download-status": {
